Storing foreign access registrations as callbacks.
